+++ laktory/models/datasources/basedatasource.py
@@ -2,10 +2,19 @@ from abc import abstractmethod
 from typing import Union
 from laktory.spark import DataFrame
 
-from laktory.models.base import BaseModel
+from laktory.models.basemodel import BaseModel
 
 
 class BaseDataSource(BaseModel):
+    """
+    Base class for building data source
+
+    Attributes
+    ----------
+    read_as_stream
+        If `True` read source as stream
+    """
+
     read_as_stream: Union[bool, None] = True
 
     @abstractmethod
@@ -13,5 +22,6 @@ class BaseDataSource(BaseModel):
         raise NotImplementedError()
 
     @property
-    def is_cdc(self):
+    def is_cdc(self) -> bool:
+        """If `True` source data is a change data capture (CDC)"""
         return getattr(self, "cdc", None) is not None